Fix #62, refactor heartbeat recipe for 0.10 release
