+++ test/z-ensure-coverage.js
@@ -11,7 +11,7 @@ require("../src/common/heartbeat/upload-validate.js");
 require("../src/common/index.js");
 require("../src/common/personinfo.js");
 require("../src/common/request.js");
-require("../src/heartbeat-by-user-first-impression/config.js");
-require("../src/heartbeat-by-user-first-impression/heartbeat-by-user-first-impression.js");
+require("../src/recipes/heartbeat-by-user-first-impression/config.js");
+require("../src/recipes/heartbeat-by-user-first-impression/index.js");
 require("../src/repairs.js");
 require("../src/runner.js");
